+++ R/print.R
@@ -81,6 +81,72 @@ plot.kardl_mplier <- function(x, variables ="all" , ...) {
 
 }
 
+#' @import stats
+#' @export
+#' @method  summary kardl_longrun
+summary.kardl_longrun <- function(object, ...) {
+  objcoef <- object$original_model$coefficients
+
+  vcov_matrix = stats::vcov( object$original_model)
+
+  my_dep <- object$depvar
+  my_indep <- object$indepvars
+
+  multipliers_coef <- object$coefficients
+
+  A <- - 1 / objcoef[my_dep]
+
+  lr_var <- sapply(names(multipliers_coef), function(i) {
+    B <- objcoef[i] / (objcoef[my_dep]^2)
+    (A^2) * vcov_matrix[i, i] +
+      2 * A * B * vcov_matrix[i, my_dep] +
+      (B^2) * vcov_matrix[my_dep, my_dep]
+  })
+
+  multipliers_coef_se <- sqrt(as.vector(lr_var))
+  tvals <- multipliers_coef / multipliers_coef_se
+  pvals <- 2 * stats::pt(-abs(tvals), df = stats::df.residual(object$original_model))
+
+  coef_table <- cbind(
+    Estimate = multipliers_coef,
+    `Std. Error` = multipliers_coef_se,
+    `t value` = tvals,
+    `Pr(>|t|)` = pvals
+  )
+
+  ans <- list(
+    call = object$call,
+    coefficients = coef_table,
+    sigma = sqrt(sum(object$residuals^2) / object$df.residual),
+    df = c(object$rank, object$df.residual, length(object$coefficients)),
+    r.squared = NA_real_,
+    adj.r.squared = NA_real_,
+    fstatistic = NA,
+    cov.unscaled = NA,
+    note = attr(object, "note"),
+    estimation_type = attr(object, "estimation_type")
+  )
+
+  class(ans) <- c("summary_kardl_longrun", "summary.lm")
+  ans
+}
+
+#' @export
+#' @method print summary_kardl_longrun
+print.summary_kardl_longrun <- function(x, digits = max(3L, getOption("digits") - 3L), ...) {
+  cat("\nCall:\n", paste(deparse(x$call), collapse = "\n"), "\n", sep = "")
+
+  cat("\nEstimation type:\n")
+  cat(x$estimation_type, "\n")
+
+  cat("\nCoefficients:\n")
+  printCoefmat(x$coefficients, digits = digits, P.values = TRUE, has.Pvalue = TRUE)
+
+  cat("\nNote:\n")
+  cat(x$note, "\n")
+
+  invisible(x)
+}
 
 #' @export
 #' @method print kardl_symmetric
@@ -105,10 +171,11 @@ print.kardl_symmetric <- function(x, ...) {
 #' @method summary kardl_symmetric
 summary.kardl_symmetric <- function(object,level=0.05 ,...) {
   decision <- list()
+  pValType<-ifelse(object$type == "Chisq","Pr(>Chisq)", "Pr(>F)")
    if(!is.null(object$Lwald)){
      decision$long_run <-list()
        for(v in rownames(object$Lwald)){
-         decision$long_run[[v]]<- ifelse( object$Lwald[v,"Pr(>F)"] < level,  paste0("Reject H0 at ",level*100,"% level. Indicating long-run asymmetry for variable ",v,"."),
+         decision$long_run[[v]]<- ifelse( object$Lwald[v,pValType] < level,  paste0("Reject H0 at ",level*100,"% level. Indicating long-run asymmetry for variable ",v,"."),
                                           paste0("Fail to Reject H0 at ",level*100,"% level. Indicating long-run symmetry for variable ",v,"."))
        }
      }
@@ -116,7 +183,7 @@ summary.kardl_symmetric <- function(object,level=0.05 ,...) {
    if(!is.null(object$Swald)){
      decision$short_run <-list()
      for(v in rownames(object$Swald)){
-       decision$short_run[[v]]<- ifelse( object$Swald[v,"Pr(>F)"] < level,  paste0("Reject H0 at ",level*100,"% level. Indicating short-run asymmetry for variable ",v,"."),
+       decision$short_run[[v]]<- ifelse( object$Swald[v,pValType] < level,  paste0("Reject H0 at ",level*100,"% level. Indicating short-run asymmetry for variable ",v,"."),
                                         paste0("Fail to Reject H0 at ",level*100,"% level. Indicating short-run symmetry for variable ",v,"."))
      }
    }
@@ -125,6 +192,7 @@ summary.kardl_symmetric <- function(object,level=0.05 ,...) {
     Swald = object$Swald,
     Lhypotheses = object$Lhypotheses,
     Shypotheses = object$Shypotheses,
+    type = object$type,
     decision = decision
   )
   class(out) <- "summary.kardl_symmetric"
@@ -134,12 +202,21 @@ summary.kardl_symmetric <- function(object,level=0.05 ,...) {
 #' @export
 #' @method print summary.kardl_symmetric
 print.summary.kardl_symmetric <- function(x, ...){
+  if(x$type == "Chisq"){
+    pValType <- "Pr(>Chisq)"
+    Valtype <- "Chi-squared"
+    testLable <- "Chisq"
+  } else if(x$type == "F"){
+    pValType <- "Pr(>F)"
+    Valtype <- "F"
+    testLable <- "F value"
+  }
 
   if (!is.null(x$Lwald) && nrow(x$Lwald) > 0) {
     cat("Long-run symmetry tests:\n\n")
     for(v in rownames(x$Lwald)){
       cat("Test for variable: ",v,"\n")
-      cat("F-value: ", x$Lwald[v,"F value"], ", p-value: ", x$Lwald[v,"Pr(>F)"], "\n")
+      cat(Valtype," statistic: ", x$Lwald[v,testLable], ", p-value: ", x$Lwald[v,pValType], "\n")
       cat("Test Decision: ", x$decision$long_run[[v]], "\n")
       cat("Hypotheses:\n")
       cat(paste0("H0: ", x$Lhypotheses$H0[[v]], "\n"))
@@ -154,7 +231,7 @@ print.summary.kardl_symmetric <- function(x, ...){
     cat("Short-run symmetry tests:\n\n")
     for(v in rownames(x$Swald)){
       cat("Test for variable: ",v,"\n")
-      cat("F-value: ", x$Swald[v,"F value"], ", p-value: ", x$Swald[v,"Pr(>F)"], "\n")
+      cat(Valtype," statistic: ", x$Swald[v,testLable], ", p-value: ", x$Swald[v,pValType], "\n")
       cat("Test Decision: ", x$decision$short_run[[v]], "\n")
       cat("Hypotheses:\n")
       cat(paste0("H0: ", x$Shypotheses$H0[[v]], "\n"))
@@ -191,7 +268,7 @@ print.summary_htest <- function(x, ...){
   cat("\nCritical Values (Case ",x$case,"):\n")
   print(x$crit_vals)
   if (!is.null(x$notes)) {
-    cat("\n\033[1;33mNotes:\033[0m\n")
+    cat("\nNotes:\n")
     for (note in x$notes) {
       cat("   \u2022 ", note, "\n", sep = "")
     }
@@ -213,7 +290,7 @@ print.kardl_long_run<- function(x, ...) {
 #' @export
 #' @method plot kardl_long_run
 plot.kardl_long_run <- function(x,  ...) {
-  cat("\033[1;33mWARNING: Residual diagnostic plots are meaningless for long-run estimators\033[0m\n")
+  cat("WARNING: Residual diagnostic plots are meaningless for long-run estimators\n")
   if (interactive()) {
     ans <- readline("Show plots anyway? (y/n): ")
     if (tolower(ans) != "y") return(invisible(NULL))
@@ -229,7 +306,7 @@ print.kardl_lm<- function(x, ...) {
   cat(paste(sprintf("%s: %d", names(x$lagInfo$OptLag), x$lagInfo$OptLag), collapse = ", "  ),"\n")
   NextMethod()  # continues with normal coefficient printing
   if (!is.null(x$notes)) {
-    cat("\n\033[1;33mNotes:\033[0m\n")
+    cat("\nNotes:\n")
     for (note in x$notes) {
       cat("   \u2022 ", note, "\n", sep = "")
     }